Improved Logger and Screenshotter
No more warnings when compiling!
The screenshot maker now returns the filename when taking a shot.
The Screenshot Maker and the GameObject Extensions now use the Logger to
write the messages
The Logger is fully documented
The logger now stores player data
The logger now prints to Debug.log, in addition to writing it to the
log.